EVAD/binary_attack_defence/requirements.txt
lief
python-magic
matplotlib
numpy
seaborn
secml
lightgbm
deap
git+https://github.com/elastic/ember.git
#torch
#torchvision